Move mock control msg to API
