+++ rust-webvr-api/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rust-webvr-api"
-version = "0.11.0"
+version = "0.11.1"
 authors = ["The Servo Project Developers"]
 
 homepage = "https://github.com/servo/rust-webvr"
